implement selection of hours.
The input hours function will get the hours from the configuration file and use what it gets to fill in the values for the hour input section.
The configuration file strings need to match exactly how it says it on the website otherwise it will not find the result.
TODO: select I certify this is accurate message and have it hit select.